Change `op` enum underlying type to `uint8_t` in `utility_linux.cpp`

+++ src/linux/utility_linux.cpp
@@ -94,7 +94,7 @@ int fileErrorToErrno(QFile::FileError error)
   }
 }
 
-enum op : unsigned int
+enum op : uint8_t
 {
   FO_COPY,
   FO_MOVE
